Clean the step regression tests up. There were several nearly identical tests that I have merged into one 2x2v step test that runs in < 1 minute including geometr initialization. I have removed the 3x2v test because it was really a pseudo-2x test with 1 cell in y. If someone wants an example for 3x SOL turbulence they should see the solovev test instead. I don't know what to do with regression tests that have reactions because I'm not sure if the user will always have ADAS installed/how that should work. I will leave STEP regression tests with reactions untouched until I know th answer to this.
